Post received new poll message in the backend to the db
